requirements.txt
transformers
bitsandbytes-cuda112
datasets
accelerate
pytorch_lightning
